Update READ.ME
For this exercise I added labels, edited scripts of the world, ball, and paddles. The game now keeps count of the lives and points the player has. The paddles are now controlled by the mouse. Overall the objective of this game is to destroy the bricks with the ball that bounces off your pad before you lose all of  your lives.